Fix GcrRegion#getRegion(String)
This modifies `GcrRegion#getRegion(String)` to take the region endpoint and
return the region, rather than taking the string equivalent of the region and
returning the region object. Based on the only way this method is used, that
was the expected behavior (and seems more useful), but not what it actually did.